ファイルの整理と失格ダイアログの追加
